+++ CHANGELOG.rst
@@ -6,6 +6,16 @@ All notable changes to this project will be documented in this file.
 The format is based on `Keep a Changelog <https://keepachangelog.com/en/1.0.0/>`_,
 and this project adheres to `Semantic Versioning <https://semver.org/spec/v2.0.0.html>`_.
 
+unreleased
+----------
+
+Added
+^^^^^
+
+* `dmarc_invalid_reports_total` metric with a count of emails from which no
+  DMARC report could be parsed.
+
+
 [0.10.1] - 2024-01-07
 ---------------------